pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-controller-prototype"
version = "0.1.0"
description = "Task-to-AgentFlow compiler prototype for coding and research workflows."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.12,<3",
  "PyYAML>=6,<7",
  "claude-code-sdk>=0.0.25,<1",
  "fastapi>=0.135,<1",
  "openhands-sdk>=1.16,<2",
  "openhands-tools>=1.16,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
